+++ docs/sidebars.js
@@ -1,6 +1,6 @@
 module.exports = {
   someSidebar: {
     "Getting Started": ['install'],
-    "Guides": ["guides/create-room", "guides/applying-room-textures", "guides/adding-objects"]
+    "Guides": ["guides/create-room", "guides/applying-room-textures", "guides/adding-objects", "guides/adding-windows", "guides/avatar-movement"]
   },
 };