+++ src/objects/furniture/FloorFurniture.tsx
@@ -6,10 +6,18 @@ import { BaseFurniture } from "./BaseFurniture";
 import { IFurniture, IFurnitureBehavior } from "./IFurniture";
 import { ObjectAnimation } from "../../util/animation/ObjectAnimation";
 import { HitEventHandler } from "../hitdetection/HitSprite";
+import { RoomPosition } from "../../types/RoomPosition";
+import { IMoveable } from "../IMoveable";
 
-export class FloorFurniture extends RoomObject implements IFurniture {
-  private baseFurniture: BaseFurniture;
-  private movementAnimation: ObjectAnimation | undefined;
+export class FloorFurniture
+  extends RoomObject
+  implements IFurniture, IMoveable {
+  private _baseFurniture: BaseFurniture;
+  private _moveAnimation: ObjectAnimation<undefined> | undefined;
+
+  private _animatedPosition: RoomPosition = { roomX: 0, roomY: 0, roomZ: 0 };
+
+  private _moving: boolean = false;
 
   private _roomX: number;
   private _roomY: number;
@@ -31,7 +39,7 @@ export class FloorFurniture extends RoomObject implements IFurniture {
 
   public set onClick(value) {
     this._onClick = value;
-    this.baseFurniture.onClick = this.onClick;
+    this._baseFurniture.onClick = this.onClick;
   }
 
   public get onDoubleClick() {
@@ -40,7 +48,7 @@ export class FloorFurniture extends RoomObject implements IFurniture {
 
   public set onDoubleClick(value) {
     this._onDoubleClick = value;
-    this.baseFurniture.onDoubleClick = this.onDoubleClick;
+    this._baseFurniture.onDoubleClick = this.onDoubleClick;
   }
 
   constructor(options: {
@@ -62,7 +70,7 @@ export class FloorFurniture extends RoomObject implements IFurniture {
     this._direction = options.direction;
     this._animation = options.animation;
 
-    this.baseFurniture = new BaseFurniture(
+    this._baseFurniture = new BaseFurniture(
       options.type,
       options.direction,
       options.animation
@@ -117,37 +125,87 @@ export class FloorFurniture extends RoomObject implements IFurniture {
   }
 
   private _updateDirection() {
-    this.baseFurniture.direction = this.direction;
+    this._baseFurniture.direction = this.direction;
+  }
+
+  private _getDisplayRoomPosition() {
+    if (this._moving) {
+      return this._animatedPosition;
+    }
+
+    return {
+      roomX: this.roomX,
+      roomY: this.roomY,
+      roomZ: this.roomZ,
+    };
   }
 
   private _updatePosition() {
-    const { x, y } = this.geometry.getPosition(
-      this.roomX,
-      this.roomY,
-      this.roomZ,
-      "object"
-    );
+    const { roomX, roomY, roomZ } = this._getDisplayRoomPosition();
 
-    this.baseFurniture.x = x;
-    this.baseFurniture.y = y;
-    this.baseFurniture.zIndex = getZOrder(this.roomX, this.roomY, this.roomZ);
+    const { x, y } = this.geometry.getPosition(roomX, roomY, roomZ, "object");
+
+    const roomXrounded = Math.round(roomX);
+    const roomYrounded = Math.round(roomY);
+    const roomZrounded = Math.round(roomZ);
+
+    this._baseFurniture.x = x;
+    this._baseFurniture.y = y;
+    this._baseFurniture.zIndex = getZOrder(
+      roomXrounded,
+      roomYrounded,
+      roomZrounded
+    );
   }
 
   private _updateAnimation() {
-    this.baseFurniture.animation = this.animation;
+    this._baseFurniture.animation = this.animation;
   }
 
-  destroy(): void {
-    this.baseFurniture.destroy();
+  move(roomX: number, roomY: number, roomZ: number) {
+    this._moveAnimation?.move(
+      { roomX: this.roomX, roomY: this.roomY, roomZ: this.roomZ },
+      { roomX, roomY, roomZ },
+      undefined
+    );
+
+    this._roomX = roomX;
+    this._roomY = roomY;
+    this._roomZ = roomZ;
+  }
+
+  clearMovement() {
+    const current = this._moveAnimation?.clear();
+
+    if (current != null) {
+      this.roomX = current.roomX;
+      this.roomY = current.roomY;
+      this.roomZ = current.roomZ;
+    }
+  }
+
+  destroyed(): void {
+    this._baseFurniture.destroy();
   }
 
   registered(): void {
-    this.roomObjectContainer.addRoomObject(this.baseFurniture);
-    this.movementAnimation = new ObjectAnimation(this.animationTicker, {
-      onStart: () => {},
-      onStop: () => {},
-      onUpdatePosition: (position, direction) => {},
-    });
+    this.roomObjectContainer.addRoomObject(this._baseFurniture);
+    this._moveAnimation = new ObjectAnimation(
+      this.animationTicker,
+      {
+        onStart: () => {
+          this._moving = true;
+        },
+        onStop: () => {
+          this._moving = false;
+        },
+        onUpdatePosition: (position, data) => {
+          this._animatedPosition = position;
+          this._updatePosition();
+        },
+      },
+      this.configuration.furnitureMovementDuration
+    );
 
     this._updatePosition();
   }